rename file "gpg" also to "bingpg" for consistency
